requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==2.5.2               # via kombu
attrs==19.3.0             # via pytest
billiard==3.6.2.0         # via celery
celery==4.4.0
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
django-cors-headers==3.2.1
django-environ==0.4.5
django==2.2.10
djangorestframework==3.11.0
idna==2.8                 # via requests
kombu==4.6.7              # via celery
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pip-tools==4.4.1
pluggy==0.13.1            # via pytest
psycopg2==2.8.4
py==1.8.1                 # via pytest
pyparsing==2.4.6          # via packaging
pytest-django==3.8.0
pytest==5.3.5
pytz==2019.3              # via celery, django
requests==2.22.0
six==1.14.0               # via packaging, pip-tools
sqlparse==0.3.0           # via django
urllib3==1.25.8           # via requests
uwsgi==2.0.18
vine==1.3.0               # via amqp, celery
wcwidth==0.1.8            # via pytest
